Rename TaskHandler to ExecutableHandler
